requirements.txt
langchain==0.1.15
python-dotenv==1.0.0
python-telegram-bot==21.6
redis==5.0.1
gigachat
pytesseract~=0.3.10
pillow~=10.3.0
requests~=2.32.3
psycopg2~=2.9.9
peewee~=3.17.3

httpx~=0.27.2
ollama~=0.3.3

langchain-community~=0.0.38
langchain-core~=0.1.52
resend~=0.8.0
langchain-text-splitters~=0.0.2